requirements.txt
streamlit==1.29.0
pandas==2.1.4
openpyxl==3.1.2
google-genai==1.30.0
python-dotenv==1.0.0
plotly==5.17.0
watchdog==6.0.0
tabulate==0.9.0